requirements.txt
flask
dnspython
python-whois
requests
ipwhois
geopy
